+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [1.4.1](https://github.com/youversion/platform-sdk-kotlin/compare/1.4.0...1.4.1) (2026-05-11)
+
+
+### Bug Fixes
+
+* **ui:** respect BibleTextOptions.textColor in BibleText and BibleIntroText YPE-2346 ([#138](https://github.com/youversion/platform-sdk-kotlin/issues/138)) ([5d5a01c](https://github.com/youversion/platform-sdk-kotlin/commit/5d5a01c0be5f9d7f43a70ecb841c91e464630a9e))
+
+# [1.4.0](https://github.com/youversion/platform-sdk-kotlin/compare/1.3.0...1.4.0) (2026-05-11)
+
+
+### Features
+
+* **ui, reader:** add search to version picker ([8315746](https://github.com/youversion/platform-sdk-kotlin/commit/83157463d7b5fcf2951228dd59822e1c3abec5df))
+
 # [1.3.0](https://github.com/youversion/platform-sdk-kotlin/compare/1.2.0...1.3.0) (2026-05-06)